# Optimize SplFixedArray dimension performance